requirements.txt

openai
supabase>=2.0.0
schedule>=1.2.0
python-dotenv>=1.0.0